Sorry for the massive commit.  It's the exception that proves the rule :)
These are the code changes for adding Date support.

Previous commits have updated the DB's to support Date fields.  This commit provides
both backend and UI support for Dates, including searching by date, adding and modifying
date fields, etc.  This change also exposed a deficiency in the internal class DataFieldValue
used for storing field values.  This was updated to properly accomodate the possibility of
a Date field, as well as future consideration for other types of data not yet supported.

Change-Id: I8671741093bbee07204aff867342a3959892b0cc