Added the foundation for providing tabular field data.  This impacted the back referencing shapes, basically exposing the fact that there was an unnecessary class used to represent back referencing data.  The code was modified to accomodate the general case, which is now seen from the user's persepctive as a Parent.  This is used by the tabular field data to reference back to the containing instance, which is why this work is done here.
The UI was also updated to provide Parent reference searching for data that has a parent.

There are additional changes in this commit having to do with license management.  This should really be
done in a separate commit.  There's no impact to the functionality right now, so these changes are just
there in preparation for a future commit that implements license management.

Change-Id: I14648882ad7f9d3c927105a5e2c0c18989649283