Modified CAL to support Query narrowing using tabular field data.  The syntax is:
    Query q;
    narrow q where {shape} : "field name" : "table column name" <operator> <value expression>

This is just an extension of the narrow syntax for non tabular fields.

Removed the use of the rich text editor in the CAL script editor dialog so that we don't have to
jump through HTML hoops in order to actually find the script content.

The CAL editor now allows an empty script, and the analytic detail handles it correctly.

Change-Id: I10d48a02fcbea7ccae292cd316ba203033215e90