Removed final keyword from CornerstoneService class so that it can be extended. Moved InstancesResponse from ChimeService so that it can be used by lower level ChimeXXXX modules.
Made the CornerstoneService ChimeConfigurable.

Added more generic support for database vendors, now including Oracle.
Created the HashTools class which contains the hashing code formerly
found inside the UserUtils Tools class.  This makes it available to
submodules like ChimeUser, as well as the Cornerstone common itself.

The DatabaseConnection class now performs a forward hash of the database password.
This is done so that the password in configuration files such as ChimeServiceFactory.xml
can not be used to gain direct access to the database.

Updaged log4j version.
Parameterized property map.