Add automagic routing of messages to their processors.
Configuring the ServiceBusMessageRouter correctly in Spring (see the Sample
project) and implementing a couple of methods on their processor will enable
users to do away hand-rolling the routing of messages to processors based
on message type and version.

If complex routing is required (routing based NOT on message type/version) then
users will still need to hand code a request message handler.

Change-Id: Ia0774d9587d7b61f6164c93e78d2881046c4fc74